L: add sym [-p] pattern
Add the sym command which plumbs or prints results for a pattern
as resolved by all configured LSPs. This enables functionality like
search or resolution of symbols from plumber. As an example, plumber
can match to java stack traces and plumb to the source file resolved
by L sym and line in the stack trace.